+++ packages/google_maps_flutter/lib/src/polyline.dart
@@ -156,6 +156,15 @@ class Polyline {
     );
   }
 
+  /// Creates a new [Polyline] object whose values are the same as this
+  /// instance.
+  Polyline clone() {
+    return copyWith(
+      patternsParam: List<PatternItem>.of(patterns),
+      pointsParam: List<LatLng>.of(points),
+    );
+  }
+
   dynamic _toJson() {
     final Map<String, dynamic> json = <String, dynamic>{};
 
@@ -234,8 +243,8 @@ Map<PolylineId, Polyline> _keyByPolylineId(Iterable<Polyline> polylines) {
     return <PolylineId, Polyline>{};
   }
   return Map<PolylineId, Polyline>.fromEntries(polylines.map(
-      (Polyline polyline) =>
-          MapEntry<PolylineId, Polyline>(polyline.polylineId, polyline)));
+      (Polyline polyline) => MapEntry<PolylineId, Polyline>(
+          polyline.polylineId, polyline.clone())));
 }
 
 List<Map<String, dynamic>> _serializePolylineSet(Set<Polyline> polylines) {