+++ packages/path_provider/example/test_driver/path_provider.dart
@@ -8,7 +8,6 @@ import 'dart:io';
 import 'package:flutter_driver/driver_extension.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:uuid/uuid.dart';
 
 void main() {
   final Completer<String> allTestsCompleter = Completer<String>();
@@ -17,43 +16,23 @@ void main() {
 
   test('getTemporaryDirectory', () async {
     final Directory result = await getTemporaryDirectory();
-    final String uuid = Uuid().v1();
-    final File file = File('${result.path}/$uuid.txt');
-    file.writeAsStringSync('Hello world!');
-    expect(file.readAsStringSync(), 'Hello world!');
-    expect(result.listSync(), isNotEmpty);
-    file.deleteSync();
+    _verifySampleFile(result, 'temporaryDirectory');
   });
 
   test('getApplicationDocumentsDirectory', () async {
     final Directory result = await getApplicationDocumentsDirectory();
-    final String uuid = Uuid().v1();
-    final File file = File('${result.path}/$uuid.txt');
-    file.writeAsStringSync('Hello world!');
-    expect(file.readAsStringSync(), 'Hello world!');
-    expect(result.listSync(), isNotEmpty);
-    file.deleteSync();
+    _verifySampleFile(result, 'applicationDocuments');
   });
 
   test('getApplicationSupportDirectory', () async {
     final Directory result = await getApplicationSupportDirectory();
-    final String uuid = Uuid().v1();
-    final File file = File('${result.path}/$uuid.txt');
-    file.writeAsStringSync('Hello world!');
-    expect(file.readAsStringSync(), 'Hello world!');
-    expect(result.listSync(), isNotEmpty);
-    file.deleteSync();
+    _verifySampleFile(result, 'applicationSupport');
   });
 
   test('getLibraryDirectory', () async {
     if (Platform.isIOS) {
       final Directory result = await getLibraryDirectory();
-      final String uuid = Uuid().v1();
-      final File file = File('${result.path}/$uuid.txt');
-      file.writeAsStringSync('Hello world!');
-      expect(file.readAsStringSync(), 'Hello world!');
-      expect(result.listSync(), isNotEmpty);
-      file.deleteSync();
+      _verifySampleFile(result, 'library');
     } else if (Platform.isAndroid) {
       final Future<Directory> result = getLibraryDirectory();
       expect(result, throwsA(isInstanceOf<UnsupportedError>()));
@@ -66,12 +45,62 @@ void main() {
       expect(result, throwsA(isInstanceOf<UnsupportedError>()));
     } else if (Platform.isAndroid) {
       final Directory result = await getExternalStorageDirectory();
-      final String uuid = Uuid().v1();
-      final File file = File('${result.path}/$uuid.txt');
-      file.writeAsStringSync('Hello world!');
-      expect(file.readAsStringSync(), 'Hello world!');
-      expect(result.listSync(), isNotEmpty);
-      file.deleteSync();
+      _verifySampleFile(result, 'externalStorage');
     }
   });
+
+  test('getExternalCacheDirectories', () async {
+    if (Platform.isIOS) {
+      final Future<List<Directory>> result = getExternalCacheDirectories();
+      expect(result, throwsA(isInstanceOf<UnsupportedError>()));
+    } else if (Platform.isAndroid) {
+      final List<Directory> directories = await getExternalCacheDirectories();
+      for (Directory result in directories) {
+        _verifySampleFile(result, 'externalCache');
+      }
+    }
+  });
+
+  final List<StorageDirectory> _allDirs = <StorageDirectory>[
+    null,
+    StorageDirectory.music,
+    StorageDirectory.podcasts,
+    StorageDirectory.ringtones,
+    StorageDirectory.alarms,
+    StorageDirectory.notifications,
+    StorageDirectory.pictures,
+    StorageDirectory.movies,
+  ];
+
+  for (StorageDirectory type in _allDirs) {
+    test('getExternalStorageDirectories (type: $type)', () async {
+      if (Platform.isIOS) {
+        final Future<List<Directory>> result =
+            getExternalStorageDirectories(type: null);
+        expect(result, throwsA(isInstanceOf<UnsupportedError>()));
+      } else if (Platform.isAndroid) {
+        final List<Directory> directories =
+            await getExternalStorageDirectories(type: type);
+        for (Directory result in directories) {
+          _verifySampleFile(result, '$type');
+        }
+      }
+    });
+  }
+}
+
+/// Verify a file called [name] in [directory] by recreating it with test
+/// contents when necessary.
+void _verifySampleFile(Directory directory, String name) {
+  final File file = File('${directory.path}/$name');
+
+  if (file.existsSync()) {
+    file.deleteSync();
+    expect(file.existsSync(), isFalse);
+  }
+
+  file.writeAsStringSync('Hello world!');
+  expect(file.readAsStringSync(), 'Hello world!');
+  expect(directory.listSync(), isNotEmpty);
+  file.deleteSync();
 }