+++ packages/device_info/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.4.1
+
+* Support the v2 Android embedding.
+* Update to AndroidX.
+* Migrate to using the new e2e test binding.
+* Add a e2e test.
+
+
+## 0.4.0+4
+
+* Define clang module for iOS.
+
 ## 0.4.0+3
 
 * Update and migrate iOS example project.